Update Project 3 final notebook.Rmd
My hypothesis is if PIK3CA is highly conserved across species, then it is likely that the amino acid sequence is similar between electric fish and mammals. Here I perform a multiple sequence alignment and create a heat plot in R studio in order to test the hypothesis. The FASTA files used for this analysis can be found under the read me file in this repository.